move DiffUtil.callback out of adapter
