+++ app/src/main/java/com/kunminx/puremusic/ui/page/adapter/PlaylistAdapter.java
@@ -19,8 +19,6 @@ package com.kunminx.puremusic.ui.page.adapter;
 import android.content.Context;
 import android.graphics.Color;
 
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.kunminx.architecture.ui.adapter.SimpleDataBindingAdapter;
@@ -35,17 +33,7 @@ import com.kunminx.puremusic.player.PlayerManager;
 public class PlaylistAdapter extends SimpleDataBindingAdapter<TestAlbum.TestMusic, AdapterPlayItemBinding> {
 
     public PlaylistAdapter(Context context) {
-        super(context, R.layout.adapter_play_item, new DiffUtil.ItemCallback<TestAlbum.TestMusic>() {
-            @Override
-            public boolean areItemsTheSame(@NonNull TestAlbum.TestMusic oldItem, @NonNull TestAlbum.TestMusic newItem) {
-                return oldItem.equals(newItem);
-            }
-
-            @Override
-            public boolean areContentsTheSame(@NonNull TestAlbum.TestMusic oldItem, @NonNull TestAlbum.TestMusic newItem) {
-                return oldItem.getMusicId().equals(newItem.getMusicId());
-            }
-        });
+        super(context, R.layout.adapter_play_item, DiffUtils.getInstance().getTestMusicItemCallback());
 
         setOnItemClickListener(((item, position) -> {
             PlayerManager.getInstance().playAudio(position);
